Revert "Updates the timestamp print format"
This reverts commit 960a9d0f2fc99244de51f14c60978b7d10bef055.

+++ bundles/pubsub/pubsub_utils/gtest/src/PubSubSerializationProviderTestSuite.cc
@@ -34,7 +34,7 @@ public:
         //Cannot reproduce this on host, so testing if a timestamp improves stability.
         const auto t = std::chrono::system_clock::now();
         char *cache = nullptr;
-        asprintf(&cache, ".pubsub_serialization_provider_cache-%lli", (long long int)t.time_since_epoch().count());
+        asprintf(&cache, ".pubsub_serialization_provider_cache-%li", t.time_since_epoch().count());
         celix_properties_set(props, OSGI_FRAMEWORK_FRAMEWORK_STORAGE, cache);
         free(cache);
